requirements.txt
ultralytics
gradio
opencv-python
pillow
numpy
pandas
openpyxl
reportlab
plotly
requests

# dev/test
pytest
